+++ script/seed.js
@@ -1,35 +1,51 @@
-'use strict'
+'use strict';
 
-const db = require('../server/db')
-const {User} = require('../server/db/models')
+const db = require('../server/db');
+const { User } = require('../server/db/models');
+const { green, red } = require('chalk');
+const faker = require('faker');
+
+//creates an array to seed the database with faker data
+let usersArr = [];
+for (let i = 1; i < 101; i++) {
+  let newUser = {
+    email: faker.internet.email(),
+    password: faker.internet.password(),
+    salt: faker.lorem.word(),
+    googleId: faker.name.firstName()
+  };
+  usersArr.push(newUser);
+}
 
 async function seed() {
-  await db.sync({force: true})
-  console.log('db synced!')
+  await db.sync({ force: true });
+  console.log(green('db synced!'));
 
-  const users = await Promise.all([
-    User.create({email: 'cody@email.com', password: '123'}),
-    User.create({email: 'murphy@email.com', password: '123'})
-  ])
+  const users = await Promise.all(
+    usersArr.map(user => {
+      return User.create(user);
+    })
+  );
 
-  console.log(`seeded ${users.length} users`)
-  console.log(`seeded successfully`)
+  console.log(green(`seeded ${users.length} users`));
+  console.log(green(`seeded successfully`));
 }
 
 // We've separated the `seed` function from the `runSeed` function.
 // This way we can isolate the error handling and exit trapping.
 // The `seed` function is concerned only with modifying the database.
 async function runSeed() {
-  console.log('seeding...')
+  console.log('seeding...');
   try {
-    await seed()
+    await seed();
   } catch (err) {
-    console.error(err)
-    process.exitCode = 1
+    console.error(red('Oh noes! That was weird...'));
+    console.error(err);
+    process.exitCode = 1;
   } finally {
-    console.log('closing db connection')
-    await db.close()
-    console.log('db connection closed')
+    console.log('closing db connection');
+    await db.close();
+    console.log('db connection closed');
   }
 }
 
@@ -37,8 +53,8 @@ async function runSeed() {
 // `Async` functions always return a promise, so we can use `catch` to handle
 // any errors that might occur inside of `seed`.
 if (module === require.main) {
-  runSeed()
+  runSeed();
 }
 
 // we export the seed function for testing purposes (see `./seed.spec.js`)
-module.exports = seed
+module.exports = seed;
